requirements.txt
audioread==3.0.1
CacheControl==0.13.1
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
contourpy==1.2.0
cryptography==41.0.7
cycler==0.12.1
decorator==5.1.1
firebase-admin==6.3.0
fonttools==4.46.0
future==0.18.3
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.25.1
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.13.0
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
httplib2==0.22.0
idna==3.6
joblib==1.3.2
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
matplotlib==3.8.2
mir-eval==0.7
msgpack==1.0.7
numba==0.58.1
numpy==1.26.2
packaging==23.2
Pillow==10.1.0
platformdirs==4.1.0
pooch==1.8.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
requests==2.31.0
rsa==4.9
scikit-learn==1.3.2
scipy==1.11.4
six==1.16.0
soundfile==0.12.1
soxr==0.3.7
threadpoolctl==3.2.0
typing_extensions==4.8.0
uritemplate==4.1.1
urllib3==2.1.0
